[#1279] Implement new landing page (#14)
* Use HTTPS version of API by default

next.openspending.org is configured to redirect automatically to the HTTPS
version. This cause that if, for example, we're developing locally on HTTP, the
API calls will try to reach the HTTP version of the API, and will get
redirected to HTTPS, ultimatelly leading to a CORS issue. This patch fixes
that.

* Upgrade Webpack to 2.6.1

* Remove "public/" folder

There's no need to keep it in source control, as all of its contents are built
from the source files.

* Load config using require() instead of global variable

I had to reconfigure the test runner to use `mocha-webpack`, otherwise the
tests would be unable to run.

* Run server on "npm start", instead of simply opening index.html

* Move HTML files to app/, and serve only files in public/

This guarantees that we won't serve files by mistake.

* Remove osExplorerService

This service was used only to load the config. As we're loading it using
webpack's `require()`, it isn't needed anymore.

* [#1279] Add Nunjucks precompiling the templates

It'll be useful when we start having multiple pages.

openspending/openspending#1279

* [#1279] Move search page to /search

The landing page will live in /. I had to fix all paths, as they're relative to
the HTML page's path. It feels ugly, but as we'll have only 2 pages, it's OK.

openspending/openspending#1279

* [#1279] Move search page to /search and add landing page on /

There're still a fex issues to be fixed, namely:

* The search input doesn't work
* Remove featured data packages
* Remove unused CSS
* Organize loading of os-bootstrap (maybe we won't need it).
* Implement user login in header
* Implement the description of numbers of datasets

And possibly others. Consider this commit just as a initial change.

openspending/openspending#1279

* [#1279] Add search query to the URL bar

Now both the filters and the search query will be added to the URL like
`/search?q=boost`, `/search?q=boost&region=sa`, `/search?viewAll`. This not only
allows users to share the actual search results page, but also allows the search
bar in the landing page to work.

openspending/openspending#1279

* [#1279] Extract Login menu item to a directive and add it to landing page

The styling is off, but the functionality works.

openspending/openspending#1279

* [#1279] Fix lint issues

openspending/openspending#1279

* [#1279] Convert Less to Sass, as our other stylesheets use SASS

openspending/openspending#1279

* [#1279] WIP Use "os-styles" for styles

Missing some cleanings, removing unused stuff. Reorganize things, and run
npm-shrinkwrap again.

openspending/openspending#1279

* [#1279] Rename "landing" to "home"

openspending/openspending#1279

* [#1279] Copy `os-styles` assets into public folder

We need them, otherwise the stylesheets won't work properly (not linking
correctly to fonts and icons, for example)

openspending/openspending#1279

* [#1279] Use SCSS instead of Gulp to create vendor CSS file

It's easier to maintain IMHO, as it's consistent with how we build all of our
stylesheet files.

openspending/openspending#1279

* [#1279] Fix styling of home page

openspending/openspending#1279

* [#1279] Fix styling of LoginMenuItem directive

I had to change the DOM structure because its styling, defined on
`os-bootstrap`, requires a specific structure.

openspending/openspending#1279

* [#1279] Don't `imagemin()` the `os-styles` assets

There are more than images in the assets. In the future, this minification
should be done on `os-styles` itself, so we avoid having to optimize assets in
every other repository.

openspending/openspending#1279

* [#1279] Load stats from the API

This uses os-conductor's new `/package/stats` endpoint.

openspending/openspending#1279

* [#1279] Remove featured data packages section

We don't have featured data packages yet. Whenever we have them, we can revert
this commit.

openspending/openspending#1279

* [#1279] Update npm-shrinkwrap

openspending/openspending#1279

* [#1279] Update code to work with themes on "os-styles"

Added via https://github.com/openspending/os-styles/pull/1

openspending/openspending#1279

* [#1279] Fix Dockerfile

Before the Dockerfile always installed from `master`, now it gets the code from
the local filesystem. I also had to add a few dependencies, mainly to install
`imagemin`.

openspending/openspending#1279

* Test our Docker image build on Travis

This ensures that we're at least able to build the image.

* Fix Dockerfile - re-add git

* Fix startup.sh